[merge/#6] login controller추가, login.jsp 파일추가

+++ whole/src/main/java/model/entity/UserEntity.java
@@ -152,7 +152,7 @@ public class UserEntity {
     }
     
 
-    public boolean isSameUser(int userid) {
+    public boolean isSameUser(long userid) {
         return this.userId == userid;
     }
